requirements.txt
pandas>=2.0.0
aiohttp>=3.9.0
openai>=1.0.0
rapidfuzz>=3.0.0
tqdm>=4.66.0
tenacity>=8.2.0
openpyxl>=3.1.0
python-dotenv>=1.0.0

# Web UI dependencies
streamlit>=1.29.0
bcrypt>=4.0.0
